deprecate pypistats download
